A small attempt of GPU memory optimization gridnet_models.py
added parameters atonce_count_limit=None, device="cpu", delay_sending_to_device=True,  to GridNetHexMM on the hopes of delaying sending ton GPU when the counts and images are full size. Recently, f_count is the logits and f_img is a 1024 vector of UNI instead of the image itself.